for miguel to resolve, animation refactored

+++ src/backEnd/GameData/State/ComponentGraphImpl.java
@@ -95,7 +95,7 @@ public class ComponentGraphImpl extends Observable implements ComponentGraph {
 			currList= new ArrayList<Component>();
 		}
 		currList.add(newComponent);
-componentMap.put(location, currList);
+		componentMap.put(location, currList);
 
 		this.setChanged();
 		this.notifyObservers();
@@ -120,7 +120,7 @@ componentMap.put(location, currList);
 	}
 	
 	@Override
-	public List<Component> getComponentsWithinRadius(Component centerComp, float radius){
+	public List<Component> getComponentsWithinRadius(Component centerComp, double radius){
 		Point2D centerLoc = (Point2D) centerComp.getAttribute("Position").getValue();
 		ArrayList<Component> componentsWithinRadius = new ArrayList<Component>();
 		for (Point2D loc : componentMap.keySet()){
